improvement(doc): codeBlock structure
