+++ packages/doc/content/components/components/button/default-native.mdx
@@ -5,33 +5,12 @@ You can use all differente compounds like `<Button.Outline />` and `<Button.Text
 
 ### Usage
 
-```javascript expo=true center=true
-<Button>Find an activity</Button>
+```javascript reactLive center=true
+<Button>
+  <Star />
+  Find an activity
+</Button>
 ```
-
-<SnackEmbed
-  dependencies="@gympass/yoga"
-  code={`import * as React from 'react';
-import { Text, View, StyleSheet } from 'react-native';
-import { ThemeProvider, Button } from '@gympass/yoga';
-const App = () => (
-  <ThemeProvider>
-    <View style={styles.container}>
-      <Button>Gympass</Button>
-    </View>
-  </ThemeProvider>
-);
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#ecf0f1',
-  }
-});
-export default App;
-`}
-/>
 
 ### Props
