+++ packages/doc/content/components/components/button/default-web.mdx
@@ -5,9 +5,10 @@ You can use all differente compounds like `<Button.Outline />` and `<Button.Text
 
 ### Usage
 
+<!--
 ```javascript reactLive=true center=true
 <Button>Find an activity</Button>
-```
+``` -->
 
 ### Props
