+++ packages/doc/src/components/CodeBlock/Expo.jsx
@@ -1,0 +1,5 @@
+const Expo = () => {
+  return 'a';
+};
+
+export default Expo;
