+++ packages/doc/src/components/CodeBlock/index.js
@@ -1,3 +1,6 @@
 import CodeBlock from './CodeBlock';
+import PrismHighlight from './PrismHighlight';
+import ReactLive from './ReactLive';
+import Expo from './Expo';
 
-export default CodeBlock;
+export { CodeBlock, PrismHighlight, ReactLive, Expo };
